Update FreeIPA Server & Client Setup
FreeIPA is an open-source identity management and security solution that is primarily designed for managing the identity and authentication of users and systems in a networked environment. It provides centralized authentication, authorization, and account information by integrating several key components. FreeIPA is commonly used in Linux and Unix environments and is similar in purpose to Microsoft Active Directory in the Windows world.